Automate the release of new versions
Whenever a new entry is made in the CHANGELOG.md file, this indicates we
need to do a new release. When the change is merged back to the default
branch, we will extract the version and list of changes from the CHANGELOG.md
file and automatically generate a new release.