Add support for using latest image
Previously the image name to use was mandatory and there was no way to
create an MBIP instance using the latest available image. This change
allows you to specify 'latest' as the image name and the terraform
module will automatically figure out what the latest image is and use
it.

The terraform openstack provider does have support for an image data
source. Unfortunately, it uses the openstack API which does not allow
for using any form of wildcard or regexp to match image names. Using the
most_recent flag also doesn't help since the VIO API implementation does
not correctly return the data needed for that sorting. To work around
these issues, we use a VIO client implemented on top of the openstack
gophercloud API client in go that allows us to sort the images by
extracting the version from the image name and allows us to filter the
list of image names using a regexp.