feat: Add support for specifying BIG-IP Next Release
The terraform module tests will now use terraform-runnver version 3.0.1
which has support for specifying the BIG-IP Next release to fetch the
latest image for as well as the changes to fetch the BIG-IP Next images
using their new name so we use the k3s images instead of the old docker
images. The tests have been updated to specify the 0.7.0 release.