feat: Add support for configuring selfIPs
Previously we were letting VIO create the network ports associated with
the internal and external network interfaces and then just configuring
our own static IP addresses on those networks since Seattle VIO allows
it. This does not work in TLV openstack. This change now creates ports
associated with the internal and external networks, assigns the
specified IP addresses to those ports, and attaches those ports to the
internal and external network interfaces. This has the added benefit of
making the IP addresses shown in the openstack UI match the IP addresses
configured on the system.