feat: Fix module to no longer be a legacy module
In order to use the count parameter on a module, it can't be a legacy
module. A legacy module is defined as a module that has the provider
configuration embedded in it. Since we no longer include the provider
configuration in this module to accomplish this, users of this module
will now have to set the provider configuration themselves.