fix: Update release process description handling
The release process was previously attempting to pass the release
description via a dotenv file. This worked fine if there was only a
single line in the release description, but caused errors uploading the
dotenv artifacts when there were multiple lines. We now use a separate
file for the release description and the release job reads directly from
that file.